streamlit_app/requirements.txt
streamlit==1.28.0
tensorflow==2.13.0
Pillow==10.0.0
matplotlib==3.7.0
numpy==1.24.0